+++ tests/unit/test_application.py
@@ -1154,7 +1154,7 @@ def test_run_error(
                 """\
                 Failed to run the build script for part 'foo'.
                 Recommended resolution: Check the build output and verify the project can work with the 'python' plugin.
-                For more information, check out: http://craft-app.com/reference/plugins.html
+                For more information, check out: http://testcraft.example/reference/plugins.html
                 Full execution log:"""
             ),
         ),
@@ -2268,9 +2268,9 @@ def test_build_planner_errors(tmp_path, monkeypatch, fake_services):
 def test_emitter_docs_url(monkeypatch, mocker, app):
     """Test that the emitter is initialized with the correct url."""
 
-    assert app.app.docs_url == "www.craft-app.com/docs/{version}"
+    assert app.app.docs_url == "www.testcraft.example/docs/{version}"
     assert app.app.version == "3.14159"
-    expected_url = "www.craft-app.com/docs/3.14159"
+    expected_url = "www.testcraft.example/docs/3.14159"
 
     spied_init = mocker.spy(emit, "init")
 
@@ -2371,3 +2371,37 @@ def test_app_config_in_help(
     expected = "app-name:  The name of the app, which is 'testcraft'."
     _, err = capsys.readouterr()
     assert expected in err
+
+
+@pytest.mark.parametrize(
+    "help_args",
+    [
+        pytest.param(["--help"], id="simple help"),
+        pytest.param(["help", "--all"], id="detailed help"),
+    ],
+)
+@pytest.mark.usefixtures("emitter")
+def test_doc_url_in_general_help(help_args, monkeypatch, capsys, app):
+    """General help messages contain a link to the documentation."""
+    monkeypatch.setattr(sys, "argv", ["testcraft", *help_args])
+
+    with pytest.raises(SystemExit):
+        app.run()
+
+    expected = "For more information about testcraft, check out: www.testcraft.example/docs/3.14159\n\n"
+    _, err = capsys.readouterr()
+    assert err.endswith(expected)
+
+
+@pytest.mark.usefixtures("emitter")
+def test_doc_url_in_command_help(monkeypatch, capsys, app):
+    """Command help messages contain a link to the command's doc page."""
+    app.add_command_group("Test", [AppConfigCommand])
+    monkeypatch.setattr(sys, "argv", ["testcraft", "app-config", "-h"])
+
+    with pytest.raises(SystemExit):
+        app.run()
+
+    expected = "For more information, check out: www.testcraft.example/docs/3.14159/reference/commands/app-config\n\n"
+    _, err = capsys.readouterr()
+    assert err.endswith(expected)